requirements.txt
Pillow
astral
atproto
emoji-country-flag
event-scheduler
geopy
pyais
pytz
timezonefinder
